Revert "history tab created "
